mclient: cursor fixes
[Bug #1]

The cursor buffer size is only set during buffer creation.
This can be problematic if mclient is started when the cursor does
not exist, e.g. if xscreensaver is running. I have seen a few occasions
where the main 24x24 cursor image is clipped.

[Solution]

We pin the cursor buffer dims to 24x24. I have empirically
found this to be the maximum cursor dims returned by the
X server.

[Bug #2]

The cursor buffer always starts off in the top left corner of
the screen and jarringly warps to the real pointer location
when the mouse is moved and the cursor is moved to the correct
location.

[Solution]

Make sure to position the cursor at the true starting position
right after creation, pretty simple.