mclient: clear damage before rendering
[Bug]

For rapid bursts of damage (like closing a small window or dragging
the mouse on the root window) we tend to miss a damage event
since we do not clear damage until we finish rendering. This
shows up as a stale buffer, i.e. after you click close on xterm
the window still shows. If you cause any other damage the system
syncs back up again.

[Solution]

Clear the damage immediately upon receipt of the event.

[Tradeoffs]

This could potentially cause cases where we re-draw the same
buffer multiple times. I tried moving it to JUST after the
XShmGetImage() call but it was still too slow.

The best solution would be to only update the exact damaged areas
on the screen. There needs to be some work done on the mflinger
side to pass back damaged regions.