Introduce PerspectiveService for perspective management
PerspectiveService is a native system service that exposes a Binder
interface for managing perspectives.

Perspectives are unique views of hardware resources--a good example
of a perspective is an OS inteface. Typically we have many OSes for
different use cases, like Linux for desktop hardware, Android for mobile
hardware, FireTV for TV hardware, etc. Now that embedded devices are getting
incredibly powerful, there is a nice opportunity to unite these different
systems as simply different views of the same hardware resources.

Perspective switches occur to best adapt a user's device to the hardware
resources on hand. Currently, we only support a desktop perspective.
When a user attaches an HDMI screen, we auto trigger the desktop
perspective. The user can then pair a bluetooth mouse and keyboard
to interact with a full desktop system.

This replaces the old perspectived C daemon. It is now implemented in
C++ in order to use the Android Binder APIs.