Add mcprepare utility for container rootfs extraction on initial boot
Clean up jessie/config by removing all of the device nodes and
the hook script that are not needed anymore. All the Android mounts
are now marked "optional", which is a custom LXC flag that will not
exit failure if the path to mount fails (e.g. it does not exist on
the host). I initially did this to get around the issue with mounting
/sdcard/DCIM/Camera...this path is only mounted after the user opens
the Camera app for the first time. To avoid tying the success of
container start up to whether some path exists on the host, I mark
all mounts as optional. Also swapped mounting /sdcard/DCIM/Camera for
mounting /sdcard/DCIM so that the mount goes into effect even if the
user has not started the Camera app, but if they DO start it after the
container is up the container still picks up the Camera folder without
having to do a container reboot.

Also add a custom boot animation (not the final one), but nice to have
something representing maru on boot for demos.