Remove HDMI auto-start logic
The new Java PerspectiveService that runs in system_server will
be responsible for detecting HDMI changes and auto-starting the
desktop. From here on out, most of the core perspective management
logic will be in system_server. This has the nice benefit of interfacing
directly with the Android UI and other services only available within
the Android Java framework.

This native PerspectiveService (rename to perspectived?) is solely
responsible for interfacing with native libraries used to
start/stop/manage perspectives.