Enable daydreaming by default
This enables daydreaming when docked or charging. The default
daydream component is DeskClock.

Interestingly, there is a bug in the default value for the default
daydream component  which prevented daydreaming from starting
by default. This overlay specifies the proper DeskClock component name.

This is super useful when using Maru Desktop because you can charge
over USB through the SlimPort cable and still interact with the desktop
while the phone is daydreaming -- no more constantly waking up the phone
or setting the sleep timeout to a large value.