Move busybox and liblxc to separate modules
This move supports including the open-source
license files along with Android's standard build.

One oddity: we include two copies of liblxc.so. One that's
liblxc.so.1 that the lxc prebuilts depend on, and liblxc-1.so
that links with perspectived. TODO: get the prebuilts to depend
on liblxc.so so we can ship a single copy.