Sync container files from /system instead of /data
This gets mcprepare ready to handle all container files being
shipped on system.img in order to avoid needing to flash
the userdata partition.

Also create the /data/maru directory in init in order to make
sure the SELinux file_context is set correctly.

(This is missing the change to device-maru.mk that moves all
the files from /data to /system because I was dumb and deleted
my old local repo and added new stuff on top)