Try to re-init an uninitialized container for each lifecycle action
Now that container files are shipped on /system and synced to /data/maru,
it is important that PerspectiveService is tolerant of the files not being
fully synced when it is spawned by Android's init.

Previously, PerspectiveService would fail fast and exit if it could not
detect a properly set up container directory. Now, PerspectiveService will
not exit immediately, but instead try to init an unitialized container for
each perspectice lifecycle action requested.

This solved the issue where PerspectiveService would immediately exit
when it happened to be started before mcprepare's syncing.