Add TimeService.apk from factory LMY48M image
For some reason, TimeService.apk is missing from the Nexus
binaries from hammerhead, even though it is mentioned in
proprietary-blobs.txt for hammerhead. We add a de-odexed version
from the trusted factory image that gets added to maru_hammerhead build.

Why TimeService.apk is important
--------------------------------

The RTC for msm8974 cannot be written to so the system always
boots with a system time in the 1970s. To get around this, Qualcomm
runs a time_daemon process that writes offsets into a persistent directory,
/data/system/time/, once the system has synced with an NTP server over
WiFi or cellular. The trigger for storing these offsets is handled by
TimeService.apk listening for TIME_SET or DATE_CHANGED Android system
broadcasts. TimeService then connects to a socket created by time_daemon
to tell it to write the newest offsets.

Without these offsets, the system always boots in the 1970s until
connected to WiFi or cellular.