Move mcprepare init service to class main
This is a quick fix that hopefully reduces the chance of
perspectived starting up on an incomplete /data/maru path that is
in the process of being copied from /system/maru on the first boot.

This is still a rare event (it can only happen on the first boot) and
seems to happen 1 out of 5 flashes.