Introduce maru's own logwrapper
logwrapper is really important to maru during the first boot of a device
to debug mcprepare as it syncs container files to /data/maru.

On Marshmallow, unlike L, init cannot execute anything without changing
domains, as seen in commit 8a0c25efb0553576afadc157b86b65eedf2ef917 of
external/sepolicy.

This is really problematic because AOSP's logwrapper is defined with the
context system_file in external/sepolicy/file_contexts, and hence denied
in the neverallow for init introduced in the above commit. The CTS can
be picky about altering things in external/sepolicy directly so it's
best to not touch that.

Instead, maru now rolls it's own logging wrapper with a specific
sepolicy. Hopefully maru gains some buffer now if newer versions of AOSP
try to pull out the rug again with sepolicy changes.

Signed-off-by: Preetam D'Souza <preetamjdsouza@gmail.com>