perspectived: Improve container management
The most important functional change here is that we interact with
liblxc in a purely stateless way--much like lxc-start and lxc-stop--by
creating a fresh lxc_container object (based on the container id, i.e.
it's name under the container directory) for every request. This ensures
that we are always using the latest config file for the container, and
most importantly, that we can recover from the race condition case where
we initialize a container off a partially synced /data/maru folder on
first boot. It also makes the race condition much less likely, since we
no longer initialize the container within PerspectiveService's
constructor--very likely to occur before mcprepare finishes--but rather
when the user actually attempts to start the container.

Note that this also paves the road for managing multiple containers in
the future, since we hand-off state management to liblxc and simply
refer to containers via their names (id).

Finally, this abstracts away container management via a common
ContainerManager interface to isolate container code to a specific
container manager class. This is way better than having a mishmash of
container management logic in PerspectiveService itself, and makes it
easier to add another container manager if needed, e.g.  chroot-based
containers for devices that lack kernel support for LXC.

Signed-off-by: Preetam D'Souza <preetamjdsouza@gmail.com>