Explicitly specify resource controllers to set up for LXC
Without this file, LXC will by default set up all resource controllers
available on the running kernel for all spawned containers, typically
cpuacct, cpuset, devices, freezer, and memory.

In the case of cpuset, this caused the unexpected limiting of all
container processes to a single CPU because of how cpusets behave with
dynamic hotplugging of cores caused by the /system/bin/mpdecision daemon
on Android.

The details are a little hard to follow so these are the exact steps to
reproduce this issue:

0. The user starts up the desktop container.

1. When LXC starts up the container for the first time off a fresh boot,
it initializes all resource controllers it needs, which previously would
be all of them, including cpuset. For cpuset specifically,
/dev/cpuset/lxc will be initialized by copying over it's parent's
configuration (the default root cpuset configuration /dev/cpuset). If
the system was idle, this will likely just be "0" for cpu0, if it was busy it
could be a larger subset, possibly "0-3" for all cores.

2. When the system idles again, the root cpuset /dev/cpuset/cpus will go
back to "0" since mpdecision aggresivelly offlines cpus during idle.
cpusets have the property that child groups must be a subset of the
parent group, so LXC's cpuset will also go back to "0".

3. When the user starts actively using the desktop again, mpdecision
will see the increased CPU load and dutifully turn on the other cores
since they are needed. /dev/cpuset/cpus is immediately updated to all
cores since it always matches the available cores on the system ("0-3"
for hammerhead) but /dev/cpuset/lxc/cpus would never sync up to this new
value due to the way cpuset children work--it is still stuck at "0". In
other words, child cpusets shrink with their parents, but do not grow
with their parents. Thus all container processes are stuck on cpu0 even
though there are more cores available for work now.

The net effect of the above chain of events is that the Android side of
Maru works perfectly since it will schedule all Android tasks on all
available cores at all times regardless of hotplugging...but the desktop
container always ends up being pinned to a single core!

To avoid any of these unexpected issues, we only set up the resource
controllers we explicitly make use of, which is the devices cgroup. We
also include cpuacct since it is only used for accounting and not
resource limiting.

Signed-off-by: Preetam D'Souza <preetamjdsouza@gmail.com>