Introduce experimental desktop audio
Based on Georgian Chituc's ideas and Twaik Yont's PulsePipePlayer
prototype as decribed in this Maru OS Dev thread:

https://groups.google.com/forum/#!topic/maru-os-dev/HzqAdt2bpNw

In essence, we are introducing the maudioflinger service which creates a
named pipe on a location that is visible both from Android and the
container (/data/maru/maudio for now) that ALSA can be configured to
write to. maudioflinger reads the ALSA samples from this pipe and passes
them over to Android's sound system via OpenSLES.

This works great with aplay and mpg321 for WAVs and MP3s but is quite
laggy with Youtube on Firefox and VLC. Regardless, this is a good
start--hoping this can serve as the foundation to get audio working well
in Maru Desktop.

Signed-off-by: Preetam D'Souza <preetamjdsouza@gmail.com>