Customize apps and launcher on LineageOS
We use a phony package to remove some LineageOS packages that are
unnecessary and to save some space on /system.

We use Launcher3 instead of Trebuchet, and remove the annoying quick
search bar (QSB) with an xml overlay.

We use the AOSP-default Provision app instead of LineageSetupWizard to
do initial setup of the device on first boot.

Signed-off-by: Preetam D'Souza <preetamjdsouza@gmail.com>