fix: mcprepare: Reduce risk of incomplete installs
There is a rare case where maru files would fail to copy over completely
to /data, but the actual files themselves would still exist (they would
just be empty files). Unfortunately, on the next boot, mcprepare would
mistakenly think that everything was a success last time and do nothing
because the directories used in the checks existed. In short, our checks
to determine the last installation state were very flaky.

To reduce the risk of this, let's make it obvious when an installation
fails prematurely: just touch a .lock file before beginning the
installation and only remove it at the end if the install was
successful.

To ensure that the lock file is only removed if all cmds succeed, we add
error checking to the two main installation operations: the cp of maru
files and the rootfs extraction.

Finally, there is clean up of the initial check logic to determine
whether an installation is needed: the checks are now all consolidated
into a get_state() function and the script is structured as a more
obvious state machine. This will make it more pleasant to add
functionality if needed in the future too.

Signed-off-by: Preetam D'Souza <preetamjdsouza@gmail.com>